@@ -12,58 +12,60 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/flex-layout": "^6.0.0-beta.18",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@angular/service-worker": "^6.1.8",
-    "@ngrx/store": "^6.1.0",
-    "@ngx-translate/core": "^10.0.2",
-    "core-js": "^2.5.4",
-    "d3": "^5.7.0",
-    "devextreme": "^18.1.6",
-    "devextreme-angular": "^18.1.6",
-    "hammerjs": "^2.0.8",
-    "jssip": "^3.2.7",
-    "localforage": "^1.5.3",
-    "moment": "^2.22.2",
-    "ngx-avatar": "^3.0.2",
-    "ngx-cookie-service": "^1.0.10",
-    "ngx-image-viewer": "^1.0.12",
-    "perfect-scrollbar": "^1.4.0",
-    "prismjs": "^1.15.0",
-    "rxjs": "~6.2.0",
-    "rxjs-compat": "^6.3.2",
-    "websocket": "^1.0.25",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "7.1.4",
+    "@angular/cdk": "7.2.0",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/flex-layout": "7.0.0-beta.22",
+    "@angular/forms": "7.1.4",
+    "@angular/http": "7.1.4",
+    "@angular/material": "7.2.0",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/router": "7.1.4",
+    "@angular/service-worker": "7.1.4",
+    "@ngrx/store": "7.0.0",
+    "@ngx-translate/core": "11.0.1",
+    "core-js": "2.6.1",
+    "d3": "5.7.0",
+    "devextreme": "18.1.6",
+    "devextreme-angular": "18.1.6",
+    "hammerjs": "2.0.8",
+    "jssip": "3.2.7",
+    "localforage": "1.5.3",
+    "moment": "2.22.2",
+    "ngx-avatar": "3.0.2",
+    "ngx-cookie-service": "1.0.10",
+    "ngx-image-viewer": "1.0.12",
+    "perfect-scrollbar": "1.4.0",
+    "prismjs": "1.15.0",
+    "rxjs": "6.3.3",
+    "rxjs-compat": "6.3.3",
+    "websocket": "1.0.25",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular/cli": "~6.2.3",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "@angular-devkit/build-angular": "0.11.4",
+    "@angular/cli": "7.0.2",
+    "@angular/compiler-cli": "7.0.0",
+    "@angular/language-service": "7.0.0",
+    "@types/jasmine": "2.8.8",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "8.9.4",
+    "codelyzer": "4.5.0",
+    "jasmine-core": "2.99.1",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "3.0.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "node-sass": "^4.11.0",
+    "protractor": "5.4.0",
+    "ts-node": "7.0.0",
+    "tslint": "5.11.0",
+    "typescript": "3.1.1",
+    "webpack": "^4.28.1"
   }
 }
