@@ -1,5 +1,5 @@
 export class CONSTANT {
     public static PAGE_SIZE: number = 8;
-    public static PAGE_SIZE_FILTER: number = 20;
+    public static PAGE_SIZE_FILTER: number = 18;
     public static CAROUSEL_SIZE: number = 4;
  }
